pyproject.toml
[tool.poetry]
name = "integration-seminar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
strawberry-graphql = "^0.221.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
dependency-injector = "^4.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
